+++ archive/preprocess.h
@@ -1,24 +1,24 @@
 #pragma once
-
-#include <fstream>
 #include <iostream>
-#include <regex>
-#include <stdexcept>
-#include <string>
+#include <fstream>
 #include <vector>
+#include <string>
+#include <stdexcept>
 
 #include <curl/curl.h>
 #include <nlohmann/json.hpp>
 
+#include "db.h"
+#include "ndb.h"
 #include "api.h"
 #include "config.h"
-#include "db.h"
 #include "model.h"
 
 namespace preprocess
 {
     using json = nlohmann::json;
-    int timeToSeconds(const std::string &time);
+
+    int stations();
+    int trains();
     int main();
-    int test();
 } // namespace preprocess
