+++ include/api.h
@@ -1,9 +1,10 @@
 #pragma once
-#include <iostream>
+
 #include <fstream>
+#include <iostream>
+#include <map>
 #include <string>
 #include <vector>
-#include <map>
 
 #include <curl/curl.h>
 #include <nlohmann/json.hpp>
