+++ include/dijkstra.h
@@ -1,6 +1,7 @@
 #pragma once
-#include <vector>
+
 #include <iostream>
+#include <vector>
 
 #include "model.h"
