+++ include/main.h
@@ -1,4 +1,7 @@
 #pragma once
-#include "preprocess.h"
-#include "dijkstra.h"
+
 #include "config.h"
+#include "dijkstra.h"
+#include "preprocess.h"
+
+int main();
